requirements.txt
alembic==1.6.2
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
Flask==2.0.0
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
Mako==1.1.4
MarkupSafe==2.0.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.15
urllib3==1.26.4
Werkzeug==2.0.1
